+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityConditionFieldFactory.java
@@ -26,8 +26,8 @@ import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
 import is.codion.framework.model.AbstractForeignKeyConditionModel;
+import is.codion.framework.model.DefaultForeignKeyConditionModel;
 import is.codion.framework.model.EntitySearchModel;
-import is.codion.framework.model.ForeignKeyConditionModel;
 import is.codion.swing.common.ui.component.combobox.Completion;
 import is.codion.swing.common.ui.component.table.FilterColumnConditionPanel.FieldFactory;
 import is.codion.swing.common.ui.component.value.ComponentValue;
@@ -121,8 +121,8 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 	}
 
 	private JComponent createEqualForeignKeyField(ColumnConditionModel<? extends Attribute<?>, ?> model) {
-		if (model instanceof ForeignKeyConditionModel) {
-			EntitySearchModel searchModel = ((ForeignKeyConditionModel) model).equalSearchModel();
+		if (model instanceof DefaultForeignKeyConditionModel) {
+			EntitySearchModel searchModel = ((DefaultForeignKeyConditionModel) model).equalSearchModel();
 
 			return inputComponents.foreignKeySearchField((ForeignKey) model.columnIdentifier(), searchModel).build();
 		}