+++ changelog.md
@@ -123,6 +123,7 @@ Codion Change Log
 - ForeignKeyConditionModel.condition() added, a shorthand for adding a condition to both the IN and EQUAL search/combobox models.
 - EntityTableConditionModel.setEqualConditionValue() added.
 - DefaultForeignKeyDetailModelLink now uses the EQUAL operator in case of a single value when setting the foreign key condition.
+- ForeignKeyConditionModel renamed DefaultForeignKeyConditionModel. 
 ### is.codion.framework.servlet
 - EntityService refactored.
 ### is.codion.swing.framework.model